+++ src/main.js
@@ -1,6 +1,6 @@
 import { registerImportHooks } from './interface/import/registerHooks.js';
 
-Hooks.once('init', () => {
+Hooks.once('init', async () => {
     console.log('Obsidian Bridge | Initializing');
-    registerImportHooks();
+    await registerImportHooks();
 });